backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
PyMongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0
